+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added email validation when using insert_model_entry() crud function.
 - Added ability to add newly created models from the top admin menu dropdown right after creation without a refresh of the page.
 - Added validation for repeatable text fields for insert_model_entry() crud function.
+- Added validation for repeatable date fields for insert_model_entry() crud function.
 - Added validation for repeatable number fields for insert_model_entry() crud function.
 
 ## 0.17.0 - 2022-05-05